realizado login e nivel de acesso as páginas já em uso

+++ Projeto no spring/vl-web/src/main/java/br/fai/vl/web/controller/BibliotecarioController.java
@@ -9,6 +9,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import br.fai.vl.model.Bibliotecario;
+import br.fai.vl.web.model.Account;
 import br.fai.vl.web.service.BibliotecarioService;
 
 @Controller
@@ -18,25 +19,61 @@ public class BibliotecarioController {
 	@Autowired
 	private BibliotecarioService service;
 
+	@GetMapping("/entrar")
+	public String getLogin(final Bibliotecario bibliotecario) {
+		return "conta/login-bibliotecario";
+	}
+
+	@PostMapping("/login")
+	private String login(final Bibliotecario bibliotecario, final Model model) {
+
+		if (service.login(bibliotecario)) {
+			return "redirect:/bibliotecario/detail/" + Account.getIdUser();
+		} else {
+			return "conta/login";
+		}
+
+	}
+
 	@GetMapping("/detail/{id}")
 	private String getBibliotecarioDetail(@PathVariable final int id, final Model model) {
-		final Bibliotecario bibliotecario = service.readById(id);
 
-		model.addAttribute("tipoUsuario", "bibliotecario");
+		if (!Account.isLogin()) {
+			return "redirect:/bibliotecario/entrar";
+		} else {
+			if (Account.getPermissionLevel() >= 2) {
 
-		model.addAttribute("usuario", bibliotecario);
-		return "usuario/detail";
+				final Bibliotecario bibliotecario = service.readById(id);
+
+				model.addAttribute("tipoUsuario", "bibliotecario");
+
+				model.addAttribute("usuario", bibliotecario);
+				return "usuario/detail";
+			} else {
+				return "redirect:/bibliotecario/entrar";
+			}
+		}
 	}
 
 	@GetMapping("/edit/{id}")
 	private String getBibliotecarioEdit(@PathVariable final int id, final Model model) {
 
-		model.addAttribute("url", "/bibliotecario/update");
+		if (!Account.isLogin()) {
+			return "redirect:/bibliotecario/entrar";
+		} else {
+			if (Account.getPermissionLevel() >= 2) {
 
-		final Bibliotecario bibliotecario = service.readById(id);
-		model.addAttribute("user", bibliotecario);
+				model.addAttribute("url", "/bibliotecario/update");
 
-		return "usuario/editar-perfil";
+				final Bibliotecario bibliotecario = service.readById(id);
+				model.addAttribute("user", bibliotecario);
+
+				return "usuario/editar-perfil";
+			} else {
+				return "redirect:/bibliotecario/entrar";
+			}
+		}
+
 	}
 
 	@PostMapping("/update")
@@ -46,23 +83,37 @@ public class BibliotecarioController {
 		return "redirect:/bibliotecario/detail/" + bibliotecario.getId();
 	}
 
-	@PostMapping("/create")
-	private String create(final Bibliotecario editora, final Model model) {
+	@GetMapping("/register")
+	public String getRegister(final Model model, final Bibliotecario bibliotecario) {
+		return "conta/register-bibliotecario";
+	}
 
-		final int id = service.create(editora);
+	@PostMapping("/create")
+	private String create(final Bibliotecario bibliotecario, final Model model) {
+
+		final int id = service.create(bibliotecario);
 
 		if (id != -1) {
-			// return "redirect:/editora/list";
+			return "redirect:/bibliotecario/detail/" + id;
 		} else {
-			// return "/editora/create";
+			return "/bibliotecario/register";
 		}
-		return "";
 	}
 
 	@GetMapping("/delete/{id}")
 	private String delete(@PathVariable final int id, final Model model) {
 
-		service.delete(id);
-		return "redirect:/usuario/list";
+		if (!Account.isLogin()) {
+			return "redirect:/bibliotecario/entrar";
+		} else {
+			if (Account.getPermissionLevel() >= 2) {
+
+				service.delete(id);
+				return "redirect:/usuario/list";
+			} else {
+				return "redirect:/bibliotecario/entrar";
+			}
+		}
+
 	}
 }
