+++ Projeto no spring/vl-web/src/main/java/br/fai/vl/web/controller/EditoraController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import br.fai.vl.model.Editora;
+import br.fai.vl.web.model.Account;
 import br.fai.vl.web.service.EditoraService;
 
 @Controller
@@ -23,19 +24,37 @@ public class EditoraController {
 	@GetMapping("/list")
 	private String getEditoraList(final Model model) {
 
-		final List<Editora> editoras = service.readAll();
-		model.addAttribute("listaDeEditoras", editoras);
+		if (!Account.isLogin()) {
+			return "redirect:/bibliotecario/entrar";
+		} else {
+			System.out.println(Account.getPermissionLevel());
+			if (Account.getPermissionLevel() >= 2) {
+				final List<Editora> editoras = service.readAll();
+				model.addAttribute("listaDeEditoras", editoras);
 
-		return "editora/list";
+				return "editora/list";
+			} else {
+				return "redirect:/bibliotecario/entrar";
+			}
+		}
+
 	}
 
 	@GetMapping("/edit/{id}")
 	private String getEditoraEdit(@PathVariable final int id, final Model model) {
 
-		final Editora editora = service.readById(id);
-		model.addAttribute("editaEditora", editora);
+		if (!Account.isLogin()) {
+			return "redirect:/bibliotecario/entrar";
+		} else {
+			if (Account.getPermissionLevel() >= 2) {
+				final Editora editora = service.readById(id);
+				model.addAttribute("editaEditora", editora);
 
-		return "editora/edit";
+				return "editora/edit";
+			} else {
+				return "redirect:/bibliotecario/entrar";
+			}
+		}
 	}
 
 	@PostMapping("/update")
@@ -48,7 +67,17 @@ public class EditoraController {
 	@GetMapping("/register-editora")
 	private String getRegisterEditora(final Editora editora) {
 
-		return "editora/create";
+		if (!Account.isLogin()) {
+			return "redirect:/bibliotecario/entrar";
+		} else {
+			if (Account.getPermissionLevel() >= 2) {
+
+				return "editora/create";
+			} else {
+				return "redirect:/bibliotecario/entrar";
+			}
+		}
+
 	}
 
 	@PostMapping("/create")
@@ -66,7 +95,17 @@ public class EditoraController {
 	@GetMapping("/delete/{id}")
 	private String deleteEditora(@PathVariable final int id, final Model model) {
 
-		service.delete(id);
-		return getEditoraList(model);
+		if (!Account.isLogin()) {
+			return "redirect:/bibliotecario/entrar";
+		} else {
+			if (Account.getPermissionLevel() >= 2) {
+
+				service.delete(id);
+				return getEditoraList(model);
+			} else {
+				return "redirect:/bibliotecario/entrar";
+			}
+		}
+
 	}
 }
